backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
boto3==1.34.110
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.0
cycler==0.12.1
dnspython==2.7.0
fastapi==0.110.0
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.5
motor==3.3.2
mpmath==1.3.0
networkx==3.4
numpy==1.26.4
opencv-python==4.9.0.80
packaging==25.0
pandas==2.3.1
passlib==1.7.4
pillow==11.3.0
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.6.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-telegram-bot==20.7
pytz==2025.2
PyYAML==6.0.2
requests==2.31.0
scipy==1.15.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.36.3
sympy==1.14.0
torch==2.7.1+cpu
torchvision==0.22.1+cpu
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.173
ultralytics-thop==2.0.14
urllib3==2.5.0
uvicorn==0.27.1
Werkzeug==3.0.1